+++ test/camt/053/cash-management-end-of-day-report.test.ts
@@ -1,3 +1,4 @@
+import { CashManagementEndOfDayReport } from '../../../src/camt/053/cash-management-end-of-day-report';
 import { CashManagementEndOfDayReport } from '../../../src/camt/053/CashManagementEndOfDayReport';
 import fs from 'fs';
